refactor(components): migrate ButtonGroup to TypeScript
feature/typescript-buttongroup
